[Refactor] Changed Z3 ref counting behavior to work around Unique
The Unique/Finalizer combination doesn't work for managing the
reference counting on the AST, Params, etc. The issue is that
the order of when certain functions are called is not a "stack". The
`Context` delete operation can be called before the AST references
are all decremented. This causes a `std::abort` call from Z3.

This new method has a central object that manages all the reference
counts and then decrements them all when the context object is deleted.

This means that it will likely be best if a solution is solved, then
the context disgarded so that the GC can cleanup everything on the
next run.