[Refactor] Moved conan build into a script instead of in Make
The previous implementation was using Makefile so that the conan
build was integrated. With the new `task` dependency structure
for `slm`, I want to avoid needing the `make` dependency.

I've moved the `conan` build into a separate script that can run
from either the `Makefile` or as a `slm.toml` dependency. I plan
to also add a similar script for windows powershell.